+++ pkg/fixtures/image/files/nodejs-centos-6.json
