+++ pkg/fixtures/image/files/wordpress-debian-8.json
