+++ pkg/fixtures/image/files/wordpress-debian-ubuntu-11.json
