+++ pkg/fixtures/image/reader_test.go
@@ -1,4 +1,4 @@
-package images
+package image
 
 import (
 	"testing"
